norm: split file, change formatting
