+++ src/parser/check_pair.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   check_pair.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: jiyunpar <jiyunpar@student.42seoul.kr>     +#+  +:+       +#+        */
+/*   By: hanbkim <hanbkim@student.42seoul.kr>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2022/11/04 11:15:05 by jiyunpar          #+#    #+#             */
-/*   Updated: 2022/11/07 11:37:28by jiyunpar         ###   ########.fr       */
+/*   Created: 2022/12/20 10:54:33 by hanbkim           #+#    #+#             */
+/*   Updated: 2022/12/20 10:54:36 by hanbkim          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
