+++ src/parser/parser.c
@@ -6,7 +6,7 @@
 /*   By: jiyunpar <jiyunpar@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/14 16:20:21 by jiyunpar          #+#    #+#             */
-/*   Updated: 2022/11/17 09:03:49 by jiyunpar         ###   ########.fr       */
+/*   Updated: 2022/12/20 11:08:46 by hanbkim          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -41,13 +41,36 @@ t_field	*make_field(t_node **cur_node)
 	return (field);
 }
 
+void	hang_from_tree(t_tree *cmd_tree, t_field *field)
+{
+	int			cur_priority;
+	int			root_priority;
+	t_tree_node	*cur_tree_node;
+
+	cur_priority = ((t_token *)(field->start_ptr->content))->priority;
+	root_priority = ((t_token *)(cmd_tree->root->content))->priority;
+	cur_tree_node = make_tree_node(field);
+	if (root_priority >= cur_priority)
+	{
+		push_left(cur_tree_node, cmd_tree->root, cmd_tree);
+		cmd_tree->root = cur_tree_node;
+	}
+	if (root_priority < cur_priority)
+	{
+		if (cur_priority != 2)
+		{
+			push_left(cur_tree_node, cmd_tree->root->right, cmd_tree);
+			push_right(cmd_tree->root->right, cur_tree_node, cmd_tree);
+		}
+		else
+			push_right_endpoint(cmd_tree->root, cur_tree_node, cmd_tree);
+	}
+}
+
 bool	parser(t_tree *cmd_tree, t_list *cmd_list)
 {
 	t_field		*field;
 	t_node		*cur_node;
-	t_tree_node	*cur_tree_node;
-	int			cur_priority;
-	int			root_priority;
 
 	cur_node = cmd_list->head;
 	field = make_field(&cur_node);
@@ -55,24 +78,7 @@ bool	parser(t_tree *cmd_tree, t_list *cmd_list)
 	while (cur_node)
 	{
 		field = make_field(&cur_node);
-		cur_priority = ((t_token *)(field->start_ptr->content))->priority;
-		root_priority = ((t_token *)(cmd_tree->root->content))->priority;
-		cur_tree_node = make_tree_node(field);
-		if (root_priority >= cur_priority)
-		{
-			push_left(cur_tree_node, cmd_tree->root, cmd_tree);
-			cmd_tree->root = cur_tree_node;
-		}
-		if (root_priority < cur_priority)
-		{
-			if (cur_priority != 2)
-			{
-				push_left(cur_tree_node, cmd_tree->root->right, cmd_tree);
-				push_right(cmd_tree->root->right, cur_tree_node, cmd_tree);
-			}
-			else
-				push_right_endpoint(cmd_tree->root, cur_tree_node, cmd_tree);
-		}
+		hang_from_tree(cmd_tree, field);
 	}
 	return (true);
 }
